core: Move utility functions to header file
